mapping widget
split off mapping panel to keep main from cluttering; result mapping to be added; image  to extraction panel
