+++ papylio/gui/main.py
@@ -9,8 +9,11 @@ from PySide2.QtGui import QStandardItem, QStandardItemModel, QIcon
 from PySide2.QtCore import Qt
 
 import matplotlib as mpl
+
+
 from matplotlib.backends.backend_qtagg import (
     FigureCanvas, NavigationToolbar2QT as NavigationToolbar)
+
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -18,6 +21,7 @@ import papylio as pp
 from papylio import Experiment, File
 from papylio.trace_plot import TracePlotWindow
 from papylio.gui.selection_widget import SelectionWidget
+from papylio.gui.mapping_widget import MappingWidget
 
 # class TreeNode:
 #     def __init__(self, node_object, parent=None):
@@ -254,7 +258,8 @@ class MainWindow(QMainWindow):
         self.update = True
 
         self.model.itemChanged.connect(self.onItemChange)
-        #imagery
+
+        # imagery
         self.image_canvas = ImageCanvas(self, width=5, height=4, dpi=100)
 
         # Create toolbar, passing canvas as first parament, parent (self, the MainWindow) as second.
@@ -267,75 +272,6 @@ class MainWindow(QMainWindow):
         self.image = QWidget()
         self.image.setLayout(image_layout)
 
-        #map settings
-        #mapping buttons definition:
-        self.button_map_label = QLabel("method:")
-        self.button_map_method_combobox = QComboBox()
-        self.button_map_options = ['icp', 'nn', 'foo']
-        self.button_map_method_combobox.addItems(self.button_map_options)
-
-
-        button_map_dist_treshold = QLineEdit()
-        button_map_dist_treshold.setPlaceholderText("dist_treshold")
-        button_map_margin = QLineEdit()
-        button_map_margin.setPlaceholderText("edge_margin")
-        #donor mapping:
-        button_map_donor_label = QLabel("donor_pks")
-        button_map_donor_method_combobox = QComboBox()
-        button_map_donor_options = ['local-maximum-auto', 'other']
-        button_map_donor_method_combobox.addItems(button_map_donor_options)
-        button_map_donor_fract_diff = QLineEdit()
-        button_map_donor_fract_diff.setPlaceholderText("fract._diff")
-        button_map_donor_ns_min = QLineEdit()
-        button_map_donor_ns_min.setPlaceholderText("nbh_min")
-        button_map_donor_ns_max = QLineEdit()
-        button_map_donor_ns_max.setPlaceholderText("nbh_max")
-        # acceptor mapping:
-        button_map_acceptor_label = QLabel("acceptor_pks")
-        button_map_acceptor_method_combobox = QComboBox()
-        button_map_acceptor_options = ['local-maximum-auto', 'other']
-        button_map_acceptor_method_combobox.addItems(button_map_acceptor_options)
-        button_map_acceptor_fract_diff = QLineEdit()
-        button_map_acceptor_fract_diff.setPlaceholderText("fraction_diff")
-        button_map_acceptor_ns_min = QLineEdit()
-        button_map_acceptor_ns_min.setPlaceholderText("filt_nbh_min")
-        button_map_acceptor_ns_max = QLineEdit()
-        button_map_acceptor_ns_max.setPlaceholderText("filt_nbh_max")
-        #mapping grid layout:
-        map_buttons_layout = QGridLayout()
-        map_buttons_layout.addWidget(self.button_map_label, 0, 0)
-        map_buttons_layout.addWidget(self.button_map_method_combobox, 0, 1)
-        map_buttons_layout.addWidget(button_map_dist_treshold, 0, 2)
-        map_buttons_layout.addWidget(button_map_margin, 0, 3)
-        map_buttons_layout.addWidget(button_map_donor_label, 1, 0)
-        map_buttons_layout.addWidget(button_map_donor_method_combobox, 1, 1)
-        map_buttons_layout.addWidget(button_map_donor_fract_diff, 1, 2)
-        map_buttons_layout.addWidget(button_map_donor_ns_min, 1, 3)
-        map_buttons_layout.addWidget(button_map_donor_ns_max, 1, 4)
-        map_buttons_layout.addWidget(button_map_acceptor_label, 2, 0)
-        map_buttons_layout.addWidget(button_map_acceptor_method_combobox, 2, 1)
-        map_buttons_layout.addWidget(button_map_acceptor_fract_diff, 2, 2)
-        map_buttons_layout.addWidget(button_map_acceptor_ns_min, 2, 3)
-        map_buttons_layout.addWidget(button_map_acceptor_ns_max, 2, 4)
-        #build:
-        map_buttons = QWidget()
-        map_buttons.setLayout(map_buttons_layout)
-        #extra:
-        perform_mapping_button = QPushButton('Perform mapping')
-        perform_mapping_button.clicked.connect(self.perform_mapping)
-        #collect:
-        map_controls_layout = QGridLayout()
-        map_controls_layout.setAlignment(Qt.AlignTop)
-        map_controls_layout.addWidget(map_buttons)
-        map_controls_layout.addWidget(perform_mapping_button, 1, 0, 1, 2)
-        #pack in widget:
-        self.map_controls = QWidget()
-        self.map_controls.setLayout(map_controls_layout)
-        self.map_controls.setMinimumWidth(150)
-        #add to tab:
-        mapping_tab_layout = QHBoxLayout()
-        mapping_tab_layout.addWidget(self.map_controls)
-        mapping_tab_layout.addWidget(self.image)
 
         #extraction--------------------------------------
         # molecules: spot detection and extraction----------------------------------------------------
@@ -442,6 +378,7 @@ class MainWindow(QMainWindow):
         #add to tab:
         extraction_tab_layout = QHBoxLayout()
         extraction_tab_layout.addWidget(self.extraction_controls)
+        extraction_tab_layout.addWidget(self.image)
 
         # self.selection = QTableWidget()
         # self.selection.setRowCount(5)
@@ -452,9 +389,8 @@ class MainWindow(QMainWindow):
         tabs.setMovable(False)
         tabs.setDocumentMode(True)
 
-        tab1 = QWidget(self)
-        tab1.setLayout(mapping_tab_layout)
-        tabs.addTab(tab1, 'Mapping')
+        self.mapping=MappingWidget(parent=self)
+        tabs.addTab(self.mapping, 'Mapping')
         tab2 = QWidget(self)
         tab2.setLayout(extraction_tab_layout)
         tabs.addTab(tab2, 'Extraction')
@@ -502,20 +438,6 @@ class MainWindow(QMainWindow):
         input = int(input)
         self.experiment.configuration['find_coordinates']['peak_finding']['minimum_intensity_difference'] = input
         self.experiment.configuration.save()
-
-    def perform_mapping(self, t):
-        print(t)
-        selected_files = self.experiment.selectedFiles
-
-        #jk-read in a default config and change one value:
-        panel_config= self.experiment.configuration['mapping']
-        panel_method = self.button_map_method_combobox.currentText()
-        panel_config['method']=panel_method
-
-        if selected_files:
-            selected_files.serial.perform_mapping(**panel_config)
-            self.image_canvas.refresh()
-            plt.show()
 
     def find_coordinates(self):
         selected_files = self.experiment.selectedFiles
